CORSAIR-40
Particles that change process are now first copied to remote cells'
particle lists. Before sending data they are copied to an outbound
buffer, which is then sent to a neighbour process with a single
MPI_Isend. This speeded up propagateCell-loop in
particle_list_skeleton.h.

I tried many different versions and this was the fastest one. One
specific note was that it is not recommended to send particles from
remote cells with an MPI struct -- the creation of the struct simply 
takes too much time.