Committing updates to 'sep' project:
 - fixed bug in power law wave intensity spectrum, class
WaveEnergySpectrumPowerlaw now integrates energy correctly. Previous
implementation was missing 1/maxWavelength factor.

- added documentation to sep_distrib_wave_energy_powerlaw.* files.

- added documentation to sep_distrib_wave_energy_base_class.* files.

- added documentation to sep_operator_wave_energy_total.* files.

- removed unnecessary files sep_includes.h and initialization.cpp.

- added an Octave macro that plots analytic spectral wave intensity
for a power law distribution, this is useful for checking that
interpolated I(lambda) in simulation is correct.

- added a test case to sep_particle_scatterer.cpp which interpolates
spectral wave intensity at several points and writes it to file. This
can be compared with the values from Octave macro (above).