Created a small program for testing integer overflow
Test program is test/read_int_overflow/main.cpp which needs to be
compiled with MPI compiler and linked against the VLSV library. The
command to run it is
mpirun -n <procs> ./main.cpp <elements>

where
<procs> is the number of processes
<elements> is the number of array elements

Program attempts to write <elements>*sizeof(size_t) bytes per process
to a file called 'test_file.vlsv'. It seems to fail when
<elements>*sizeof(size_t) exceeds numeric_limits<int>::max(), despite
the fact that all values passed to MPI are much less than that.