Added another test program
This one should only be called with one MPI process. It tests a
mock-up multiwrite by creating an MPI_Struct that should write out
all data in a big array using a single MPI call. As it happens, the
write fails if the total number of bytes in the struct exceeds 2.14 GB.