Tested append mode
Works with generic unstructured mesh, should work with other cases too
since the changes are quite simple. The master process simply reads in
the existing footer and its position, and writes new data to where the
old data ended (overwriting old footer). Updated footer is written to
the end of file, as usual.